deployscript/pyproject.toml
[tool.poetry]
name = "infobob-docker-deploy"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.scripts]
infobob-docker-deploy = 'infobob_docker_deploy:main'

[tool.poetry.dependencies]
python = "^3.5"
attrs = "^18.2"
click = "^7.0"
docker = "^3.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
